remove gson adapter
