--- paper/src/main/java/cloud/grabsky/configuration/paper/adapter/ItemFlagTypeAdapter.java
@@ -1,43 +1,0 @@
-/*
- * MIT License
- *
- * Copyright (c) 2023 Grabsky <44530932+Grabsky@users.noreply.github.com>
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- *  of this software and associated documentation files (the "Software"), to deal
- *  in the Software without restriction, including without limitation the rights
- *  to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- *  copies of the Software, and to permit persons to whom the Software is
- *  furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- *  copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * HORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- */
-package cloud.grabsky.configuration.paper.adapter;
-
-import cloud.grabsky.configuration.adapter.AbstractEnumTypeAdapter;
-import org.bukkit.inventory.ItemFlag;
-
-/**
- * Converts {@link String} to {@link ItemFlag} and vice-versa but using case-insensitive strategy.
- */
-public final class ItemFlagTypeAdapter extends AbstractEnumTypeAdapter<ItemFlag> {
-
-    /**
-     * Default instance of {@link ItemFlagTypeAdapter}.
-     */
-    public static final ItemFlagTypeAdapter INSTANCE = new ItemFlagTypeAdapter();
-
-    private ItemFlagTypeAdapter() {
-        super(ItemFlag.class, false);
-    }
-
-}
